Write also Pose data to file at the end of sequence
A sequence is a recording. During the sequence, we store
all the Pose data, and write point clouds to file,
by snapshot or auto mode. At the end of the sequence,
we write the pose data to a new file

Add button "Start of Sequence"
Change button "New File Group" to "Finish Sequence"